make all sit_events eligibilty columns floats: fixes issue where a string representation of a float ie "1.0" will result in failure when calling .astype(int) on the column

+++ libcbm/input/sit/sit_format.py
@@ -290,9 +290,9 @@ def get_disturbance_eligibility_columns(index):
             eligibility columns
     """
     return [
-        {"name": "MinYearsSinceDist", "index": index + 0, "type": int},
-        {"name": "MaxYearsSinceDist", "index": index + 1, "type": int},
-        {"name": "LastDistTypeID", "index": index + 2, "type": int},
+        {"name": "MinYearsSinceDist", "index": index + 0, "type": float},
+        {"name": "MaxYearsSinceDist", "index": index + 1, "type": float},
+        {"name": "LastDistTypeID", "index": index + 2, "type": float},
         {"name": "MinTotBiomassC", "index": index + 3, "type": float},
         {"name": "MaxTotBiomassC", "index": index + 4, "type": float},
         {"name": "MinSWMerchBiomassC", "index": index + 5, "type": float},
